+++ redirect/click-aggregator-api/src/model/clickstream.rs
@@ -44,7 +44,7 @@ pub struct ClickStreamItem {
     /// First session timestamp (optional)
     pub session_first: Option<DateTime<Utc>>,
     /// Number of clicks in session (optional)
-    pub session_clicks: Option<u128>,
+    pub session_clicks: Option<u64>,
     /// Whether this is a unique click
     pub is_unique: bool,
     /// Whether this click is from a bot
